+++ src/Resources/contao/dca/tl_version.php
@@ -13,15 +13,12 @@ $arrDca = &$GLOBALS['TL_DCA']['tl_version'];
 
 $arrFields = [
     'memberusername'         => [
-        'label' => &$GLOBALS['TL_LANG']['tl_version']['memberusername'],
         'sql'   => "varchar(255) NULL"
     ],
     'memberid'               => [
-        'label' => &$GLOBALS['TL_LANG']['tl_version']['memberid'],
         'sql' => "int(10) unsigned NULL"
     ],
     'formhybrid_backend_url' => [
-        'label' => &$GLOBALS['TL_LANG']['tl_version']['formhybrid_backend_url'],
         'sql' => "varchar(255) NULL"
     ]
 ];